Fix defined?(@@cvar) to use cached scope and disably dummy scope.
With this commit, scopes that have classes, methods, constants,
and cvars no longer have a requirement that a dummy scope be
pushed. The scope is loaded from the same cache as other values,
and as a result those methods have no implicit cost over the same
code with no scope access.